Koszyk

+++ WK-master/WydarzeniaKulturalne.Data/Entities/ElementKoszyka.cs
@@ -14,8 +14,8 @@ namespace WydarzeniaKulturalne.Data.Entities
         public int IdElementuKoszyka { get; set; }
         public string IdSesjiKoszyka { get; set; }
         public int IdBilet { get; set; } //idtowaru
-        public virtual Bilety Bilety { get; set; } // public virtual Wydarzenie Wydarzenie 
-        public decimal Ilosc { get; set; }
+        public  virtual Bilety Bilety { get; set; } // public virtual Wydarzenie Wydarzenie 
+        public int Ilosc { get; set; }
         public DateTime DataUtworzenia { get; set; }
